Make it hard to create invalid rates for a service (#120)
This changes the UI so that you can only add rates for rategroups that
haven't been defined yet. Additionally, the form cannot be saved until
the RateGroupOverview is valid.

As this is lacking proper backend validation, invalid entries can still
happen - if you add a rate and then leave the page without fixing the
validation errors, there will still be a broken rate. If you somehow
manage to click the add button several times, there will be several
rates with the same rateGroup.

For normal usage, this should be good enough, however.