reload activites for each project in timetrack
context: if we fetch all activities at once, doctrine will load their associated timeslices even we excluded it in the JMS annotations. It is possible not possible in our case to make the Activity entity unaware of its relation to the timeslices nor to rewrite the all method to prefetch the relation because Doctrine and JMS would still do the same stuff as before.